+++ nodeconductor/structure/tests/urls.py
@@ -10,6 +10,7 @@ def register_in(router):
     router.register(r'test', views.TestServiceViewSet, base_name='test')
     router.register(r'test-service-project-link', views.TestServiceProjectLinkViewSet, base_name='test-spl')
     router.register(r'test-instances', views.TestInstanceViewSet, base_name='test-instances')
+    router.register(r'test-new-instances', views.TestNewInstanceViewSet, base_name='test-new-instances')
 
 
 router = DefaultRouter()